Moved cluster_toolkit patch out of modeling.py and changed it back to the intended goal. Better documented it and it is now being tested
